TINY-10490: fix test
